Moved to contract payment channels
Instead of using ethereum transactions for channel payments I've moved
to a channel method using a contract. This defeats the issue where one
could funnel payment and, before the recipient has a chance to redeem
the channeled transaction, the payer creates a minimum transaction
invalidating the recipients transaction.

The contract is instead to be used where valid signatures are given to
the contract which release the funds instead. Signatures are formed out
of SHA3(R || A) where R = recipient, A = amount.